Update part2a+.py
In Task 2, it was not clear if we need to implement specific parameters or just create a default of different menu types and adding specific parameters will be in the later stages. Therefore, I created both options:

Part 2 - Implements the default of different menu types, meaning just the framework for creating a free-text or numeric menu.

Part 2+ -  Includes the implementation of specific parameters, meaning the parameters I wrote (in Task 4) and that appear in the complete code.

In this updated version, I added more general examples instead of the examples written in the complete code.